Remove Yo from peerDependencies. Closes #118
The commit remove Yo as peer dependency of generator.
The reason of this commit is to not show warnings about generators
to end users in recent version of NPM during generator instalation.
The final decision about exact peerDependencies is not yet know, but
removing it won't hurt generator itself and left users with less warnings
and outputs from install routine(s)