:memo: Rename ASP.NET 5 to ASP.NET Core 1.0
This commit apply recent framework renaming as announced on MSDN (and
elsewhere):
https://blogs.msdn.microsoft.com/webdev/2016/01/19/asp-net-5-is-dead-int
roducing-asp-net-core-1-0-and-net-core-1-0/
The changes in this commit cover only renaming in original content -
with changes in sources based on aspnet/Templates left for future
update via RC2.
The rename also was not applied to external content and articles links.
Thanks!