:art: Always create `global.json` at the project root. Closes #836 (#864)
* :art: Add global.json template file

This adds only template file that will be used by content creation
scripts. The content of *version* key will be set according to
dynamic variables constructed on the generator run

* :white_check_mark: Test coverage for global.json

This adds tests for all cases when global.json is created
by main generator. The tests are slightly too verbose and
should be later refactor to use configuration for version
of SDK value instead of using hardcoded strings.

* :art: Always create global.json at the project root

This commit actually update all main generator option scripts
to always output `global.json` file on the root of the project
in order to fix problems reported by users with applications
deployed to service like Azure web applications.
When the SDK and dotnet is more mature there should be no longer
required for such fix hopefully

The same feature could be implemented as an startup option, for
example as `--global` option, but it seemed to be more user friendly
to just output that automatically - as the entire dotnet project
concept is still under active rewrite during .csproj migration

Thanks!