Add placeholder variable for ApplicationInsights
This adds placeholder for adding option for users to decide if
ApplicationInsight should be part of created content. That
option is available - but not used - in dotnet/cli templating
and as the content will be synchronized - it will be better
to have that placeholder for someone integrating templates between
repos.

The templates will contain an <%= if(....) => block

Thanks!