Remove subgenerators that are not part of RC4. Closes #891
There are tons of subgenerators that are not offical part
fo dotnet/templating nor aspnet/templates.
As the MS has shipped RC4 alrady there is really no time to
work now with tons on not yet tested nor updated C# code for
sugenerators templates (TagHelpers anyone, Middleware) that
are not part of RC4 release nor tested udner RC4 bits.
All of these subgenerators can be reintroduced into subgenerator
at later time (if there is an effort to do this) - but at this
moment leaving them without preflight checks and updates for RC4
would results in bad experience for end users.

The efforts should be put into dotnet/templating features - because
as life teach us - the bits from dotnet/templating could land in
public first, as there is a momentum and a large organization behind
that project

Thanks!